feat: msw 관련 목데이터, msw 코드 구현 (#433)
