+++ frontend/src/mocks/browser.ts
@@ -1,24 +1,9 @@
 import { setupWorker } from 'msw';
-import {
-  errorPostLike400handlers,
-  errorPostLike403Handlers,
-  errorPostLike404Handlers,
-  errorPostLike500Handlers,
-  error401handlers,
-  handlers,
-} from './handlers';
+import { DetailPageSuccessHandler, DetailPageErrorHandler } from '~/mocks/detailPage/handler';
+import { MainPageSuccessHandler, MainPageErrorHandler } from '~/mocks/mainPage/handler';
+import { WishListPageSuccessHandler, WishListPageErrorHandler } from '~/mocks/wishListPage/handler';
 
-// 좋아요 성공 시
-export const worker = setupWorker(...handlers);
+const successHandler = [...DetailPageSuccessHandler, ...MainPageSuccessHandler, ...WishListPageSuccessHandler];
+const errorHandler = [...DetailPageErrorHandler, ...MainPageErrorHandler, ...WishListPageErrorHandler];
 
-// 400 에러
-// export const worker = setupWorker(...errorPostLike400handlers);
-
-// 403에러
-// export const worker = setupWorker(...errorPostLike403Handlers);
-
-// 404에러
-// export const worker = setupWorker(...errorPostLike404Handlers);
-
-// 500에러
-// export const worker = setupWorker(...errorPostLike500Handlers);
+export const worker = setupWorker(...successHandler);
